requirements.txt
#To activate virtual enviorment run: .venv\Scripts\activate 
#To run flask local server: run flask

Flask==3.0.2
psycopg2==2.9.9
python-dotenv==1.0.1